+++ src/widgets/header/index.ts
@@ -1,0 +1,1 @@
+export { default as Header } from './ui/index';
