+++ src/widgets/header/model/navigation.ts
